requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
ntplib==0.4.0
twilio==9.3.3
flask==3.0.3
python-dotenv==1.0.1
groq>=0.13.0
sounddevice==0.4.7
soundfile==0.12.1
numpy>=1.24.0
